requirements.txt
Django==1.6.2
argparse==1.2.1
distribute==0.6.24
django-appconf==0.6
django-compressor==1.3
django-crispy-forms==1.4.0
django-registration==1.0
djangorestframework==2.3.13
six==1.6.1
wsgiref==0.1.2
Django==1.6.2
amqp==1.4.4
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.16
celery==3.1.10
distribute==0.6.24
django-appconf==0.6
django-celery==3.1.10
django-compressor==1.3
django-crispy-forms==1.4.0
django-devserver==0.7.0
django-leaflet==0.13.0
django-registration==1.0
djangorestframework==2.3.13
kombu==3.0.14
psycopg2==2.5.2
pytz==2014.2
six==1.6.1
wsgiref==0.1.2
Django==1.6.2
amqp==1.4.4
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.16
celery==3.1.10
distribute==0.6.24
django-appconf==0.6
django-celery==3.1.10
django-compressor==1.3
django-crispy-forms==1.4.0
django-devserver==0.7.0
django-geojson==2.4.0
django-leaflet==0.13.0
django-registration==1.0
djangorestframework==2.3.13
kombu==3.0.14
psycopg2==2.5.2
pytz==2014.2
six==1.6.1
wsgiref==0.1.2
